Move `createDbConfig` out of api module, to its only use site
