+++ components/db/db-connection-manager/src/test/kotlin/DbConfigTest.kt
@@ -1,14 +1,12 @@
 import com.typesafe.config.ConfigFactory
 import com.typesafe.config.ConfigValueFactory
 import net.corda.db.connection.manager.DBConfigurationException
-import net.corda.db.connection.manager.createDbConfig
 import net.corda.db.connection.manager.createFromConfig
 import net.corda.db.core.DataSourceFactory
 import net.corda.libs.configuration.SmartConfigFactory
 import net.corda.libs.configuration.SmartConfigImpl
 import net.corda.libs.configuration.secret.SecretsLookupService
 import net.corda.schema.configuration.DatabaseConfig
-import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
 import org.mockito.kotlin.any
@@ -151,68 +149,5 @@ class DbConfigTest {
                 mock(),
                 mock()))
         }
-    }
-
-    @Test
-    fun `when createDbConfig can be read`() {
-        val createdConfig = createDbConfig(
-            smartConfigFactory,
-            user,
-            pass,
-            driver,
-            url,
-            poolsize,
-            minPoolSize,
-            idleTimeout = 0,
-            maxLifetime = 0,
-            keepaliveTime = 0,
-            validationTimeout = 0,
-            key = "database-password")
-
-        assertThat(createdConfig.getString(DatabaseConfig.DB_USER)).isEqualTo(user)
-        assertThat(createdConfig.getConfig(DatabaseConfig.DB_PASS)).isEqualTo(secretConfig)
-        assertThat(createdConfig.getString(DatabaseConfig.DB_PASS + ".secret")).isEqualTo("secret value")
-        assertThat(createdConfig.getString(DatabaseConfig.JDBC_DRIVER)).isEqualTo(driver)
-        assertThat(createdConfig.getString(DatabaseConfig.JDBC_URL)).isEqualTo(url)
-        assertThat(createdConfig.getInt(DatabaseConfig.DB_POOL_MAX_SIZE)).isEqualTo(poolsize)
-        assertThat(createdConfig.getInt(DatabaseConfig.DB_POOL_MIN_SIZE)).isEqualTo(minPoolSize)
-    }
-
-    @Test
-    fun `when createDbConfig leave default driver empty`() {
-        val createdConfig = createDbConfig(
-            smartConfigFactory,
-            user,
-            pass,
-            jdbcUrl = url,
-            maxPoolSize = poolsize,
-            minPoolSize = null,
-            idleTimeout = 0,
-            maxLifetime = 0,
-            keepaliveTime = 0,
-            validationTimeout = 0,
-            key = "database-password"
-        )
-
-        assertThat(createdConfig.hasPath(DatabaseConfig.JDBC_DRIVER)).isFalse
-    }
-
-    @Test
-    fun `when createDbConfig leave default min pool size empty`() {
-        val createdConfig = createDbConfig(
-            smartConfigFactory,
-            user,
-            pass,
-            jdbcUrl = url,
-            maxPoolSize = poolsize,
-            minPoolSize = null,
-            jdbcDriver = driver,
-            idleTimeout = 0,
-            maxLifetime = 0,
-            keepaliveTime = 0,
-            validationTimeout = 0,
-            key = "database-password")
-
-        assertThat(createdConfig.hasPath(DatabaseConfig.DB_POOL_MIN_SIZE)).isFalse
     }
 }
